First stylecop cleaning
