+++ devices/Modbus/Consts.cs
@@ -1,4 +1,7 @@
-﻿namespace Iot.Device.Modbus
+﻿// Licensed to the .NET Foundation under one or more agreements.
+// The .NET Foundation licenses this file to you under the MIT license.
+
+namespace Iot.Device.Modbus
 {
     /// <summary>
     /// Contains protocol constant definitions.
