+++ devices/Modbus/Modbus.nfproj
@@ -31,20 +31,26 @@
   <ItemGroup>
     <Compile Include="Client\ModbusClient.cs" />
     <Compile Include="Consts.cs" />
-    <Compile Include="Enums.cs" />
+    <Compile Include="Enums\FunctionCode.cs" />
+    <Compile Include="Enums\ErrorCode.cs" />
     <Compile Include="Port.cs" />
+    <Compile Include="Protocol\Protocol.cs" />
     <Compile Include="Server\ModbusServer.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="Protocol\Request.cs" />
     <Compile Include="Protocol\Response.cs" />
+    <Compile Include="Structures\HoldingRegister.cs" />
+    <Compile Include="Structures\InputRegister.cs" />
     <Compile Include="Structures\Register.cs" />
     <Compile Include="Structures\Coil.cs" />
     <Compile Include="Structures\DiscreteInput.cs" />
     <Compile Include="Structures\ModbusObject.cs" />
+    <Compile Include="Util\ByteExtension.cs" />
     <Compile Include="Util\Crc16.cs" />
-    <Compile Include="Util\Converter.cs" />
+    <Compile Include="Util\BoolConverter.cs" />
     <Compile Include="Util\DataBuffer.cs" />
     <Compile Include="Server\ModbusDevice.cs" />
+    <Compile Include="Util\Int16Converter.cs" />
   </ItemGroup>
   <ItemGroup>
     <Reference Include="mscorlib">
