+++ devices/Modbus/Structures/HoldingRegister.cs
@@ -1,0 +1,9 @@
+﻿// Licensed to the .NET Foundation under one or more agreements.
+// The .NET Foundation licenses this file to you under the MIT license.
+
+namespace Iot.Device.Modbus.Structures
+{
+    internal class HoldingRegister : Register
+    {
+    }
+}
