+++ devices/Modbus/Structures/Register.cs
@@ -1,8 +1,11 @@
-﻿using System;
+﻿// Licensed to the .NET Foundation under one or more agreements.
+// The .NET Foundation licenses this file to you under the MIT license.
+
+using System;
 
 namespace Iot.Device.Modbus.Structures
 {
-    abstract class Register : ModbusObject
+    internal abstract class Register : ModbusObject
     {
         public ushort Value
         {
@@ -10,27 +13,24 @@ namespace Iot.Device.Modbus.Structures
             {
                 var blob = new[] { HiByte, LoByte };
                 if (BitConverter.IsLittleEndian)
+                {
                     blob.JudgReverse();
+                }
 
                 return BitConverter.ToUInt16(blob, 0);
             }
+
             set
             {
                 var blob = BitConverter.GetBytes(value);
                 if (BitConverter.IsLittleEndian)
+                {
                     blob.JudgReverse();
+                }
 
                 HiByte = blob[0];
                 LoByte = blob[1];
             }
         }
-    }
-
-    class HoldingRegister : Register
-    {
-    }
-
-    class InputRegister : Register
-    {
     }
 }
