+++ devices/Modbus/samples/Modbus.Samples.nfproj
@@ -42,6 +42,7 @@
     </Reference>
   </ItemGroup>
   <ItemGroup>
+    <None Include="..\.editorconfig" Link=".editorconfig" />
     <None Include="packages.config" />
     <None Include="*.md" />
   </ItemGroup>
